Fuzzy lookup within people table

Fuzzy lookup within people table

Usage

people_search(x)

Arguments

x

[character] A character string to be searched within the names in the people table

Value

The people table subset to rows that fuzzy match x

Examples

people_search("Frank Thomas")
#> # A tibble: 8 × 5
#>   name           birth_year key_mlbam key_bbref   key_fangraphs
#>   <chr>               <int>     <int> <chr>               <int>
#> 1 Frank Thompson       1854    123281 "thompfr01"       1013006
#> 2 Frank Thompson       1895    123282 "thompfr02"       1013007
#> 3 Frank Thompson       1973    440254 ""                     NA
#> 4 Hank Thomas          2002    808810 ""                     NA
#> 5 Frank Tokash         1950    565267 ""                     NA
#> 6 Frank Thomas         1968    123245 "thomafr04"           255
#> 7 Frank Thomas         1965    565093 ""                     NA
#> 8 Frank Thomas         1929    123246 "thomafr03"       1012976
people_search("ichiro")
#> # A tibble: 39 × 5
#>    name              birth_year key_mlbam key_bbref   key_fangraphs
#>    <chr>                  <int>     <int> <chr>               <int>
#>  1 Shinichi Ohnuki         1994    692702 ""                     NA
#>  2 Rich Rowland            1964    121512 "rowlari01"       1011262
#>  3 Toshihiro Noguchi       1971    538182 ""                     NA
#>  4 Yasushi Hirose          1974    530463 ""                     NA
#>  5 Akihiro Hakumura        1991    685544 ""                     NA
#>  6 Toshihiro Iwao          1987    649127 ""                     NA
#>  7 Akihiro Maeda           1983    547669 ""                     NA
#>  8 Brett Wichrowski        2002    814489 ""                     NA
#>  9 Yoshihiro Tamura        1994    684104 ""                     NA
#> 10 Soichiro Yamazaki       1998    810077 ""                     NA
#> # ℹ 29 more rows